labs/setup: remove dist-upgrade
Patch from Thomas Petazzoni:
It is really a waste of time to ask people to do a dist-upgrade,
especially if the network is slow, the machine have never been updated
from the default install, and the machines are slow. And we don't
really need the machines to be fully updated.