Don't tell to hardcode ARCH and CROSS_COMPILE
Patch from Thomas Petazzoni

Hard-coding ARCH and CROSS_COMPILE in the Makefile is a bad practice,
leads to errors (like people doing ARCH = arm with a space at the end
of the line, which doesn't work), and is definitely not the correct
way: if you end up having your kernel versioned in Git, Git will
notice a change to the source code, etc.

So clearly, the correct ways are either in the environment of on the
command line. People should simply learn those methods, and remember
to pass the values as needed.