Convert agenda to latex
Conversion of the legacy kernel agenda.
Based on the work done by Thomas for buildroot agenda proposal.

Some change have been made to be closer to the current layout: in the
summary part the column are colored in the same way of what have been
done for the odt version.

Some improvement could be made:
- police seems to big
- itemize doesn't look very fancy

The purpose of this conversion is to have a reference for the next
version of the kernel training

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>