Update the kernel instructions
It also adds a diagram that shows the NAND flash layout with the
bootloaders and the kernel.

It deletes the useless kernel patches. This also means that the
sysdev/kernel/ directory will be empty, and therefore removed from
Git. If we want to keep those directories to let our training
participants organize their work in sub-directories, then we should
store an empty '.empty' file in Git, and make the script that creates
the felabs.tar.bz2 tarball smarter so that it removes those '.empty'
files.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>