Integrate kernel agenda to the build system
This commit fully integrates the build of the kernel agenda into the
training materials build system. Until now, it was using its own
Makefile, which depended on the 'latex' Free Electrons internal
repository. Now, everything needed to build the agendas is within the
training-materials repository. Doing 'make kernel-agenda.pdf' is
enough to build the agenda.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>