kernel: add a different chapter for kernel sources
Instead of using the same contents as the sysdev course, we use our
own content for the kernel source. It removes all the discussion on
how to use the tarballs and the patches, and explains how to get the
kernel sources with Git and create a branch from a given stable
release.

As we discussed in the past, generating and manipulating kernel
patches with just patch and diff is no longer done by anybody, and
it's not the point of a kernel development course to teach about patch
and diff.

For now, it is kept for the sysdev course, we can discuss what to do
about this later on.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>