kernel: split chapter on "kernel architecture for device drivers" in two chapters
The existing chapter "Kernel architecture for device drivers" was
covering both the device model and kernel frameworks. In the new
format of the training, we'll be covering those two topics separately:
first the device model (to detect and probe an I2C device) and then
kernel frameworks (to expose the Nunchuk functionality to userspace
through the input subsystem).

This also has the nice advantage of splitting this section, which was
pretty long for the training participants.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>