Improve kernel frameworks to introduce character drivers
We no longer detail the entire API to register raw character drivers,
since we're showing how to use kernel frameworks (input and
misc). However, to explain what a kernel framework is, what is an
ioctl, we need to introduce a few details about character drivers.

Therefore, this commit extends the "Kernel Frameworks" slides by
starting with a introduction on character drivers, with many slides
and diagrams borrowed from our previous presentation on this topic.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>