kernel-frameworks: correct net_device links
When creating net_devices, the private structure is actually allocated
at the same time as the net_device structure. This is kind of a special
case. Also when working with usb devices, You don't actually have to
keep a pointer to the struct usb_interface to be able to read/write
to/from the device so you can't go back and forth between structures. So
I added a simpler example using an spi_device registering an rtc_device
in the RTC framework which is closer to what you can usually encounter.

Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>