kernel-dma: remove non-sense comment
To prevent the CPU from accessing the DMA buffer while it is mapped,
using "locking" doesn't make any sense. It's not at all about locking,
it's simply about making sure the driver does not access it. It has
nothing to do with locking in the sense of spinlocks or mutexes.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>