kernel: slightly reorganize the intro slides + virtual fs -> pseudo fs
We more classical terminology for /proc and /sys is "pseudo
filesystems" (that's how they are named in the kernel configuration),
so we now use this terminology instead of "virtual filesystems", which
can be confusing with the VFS (virtual filesystem), that is not the
same thing.

Also re-order a bit the slides to be in a more logical order.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>