makefile: have the labs generate \training too
Up until now, the \training variable was only filled with SLIDES_TRAINING, even
though we were building a lab.

Obviously, it was resulting in a empty variable whenever in the labs.

Introduce a TRAINING variable, that will be output to \training, that will be
set either to SLIDES_TRAINING or to LABS_TRAINING, so that we can rely on it in
any case.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>