sysdev-application-debugging: use the correct sysroot
Since we are now using the Buildroot generated system in the debugging
lab, the sysroot that should be used is the one created by
Buildroot. Even though for the specific case of vista-emulator it
doesn't make much difference because it only uses the C library, it
would make a difference if the application to debug used other
libraries built by Buildroot (ncurses, alsa-lib, etc.).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>