kernel-module-simple: don't tell to use indent
Last week, at the training session, they tried to use "indent -linux",
and it actually generated some code that is not compliant according to
checkpatch.

And anyway, having to use "indent -linux" is really a sign that the
source code editor is either badly chosen or improperly configured. I
did a quick poll at the Toulouse office, and none of us have 'indent'
installed on our machine, even though we have sent many many kernel
patches.

Therefore, simply get rid of the instructions telling to use indent,
as it is not recommending a good practice. Recommend instead to
configure properly its text editor.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>