yocto: labs: Update Eclipse SDK
Update the Yocto's lab about the Eclipse's plugin.
The update contains the new version of Eclipse to get the last plugin
(luna version), adds some plugins required and removes "Hob" reference
as it is now deprecated.
Instead, we should use "toaster" but a build is not direct. The user needs
to configure a Toaster project to be able to build an image.

Maybe, we could remove it ? Added some comments in the LaTeX file about
a possible way to add Toaster and comments all the "build using Eclipse" part.
Currently, Eclipse is used only to get the project view and create a recipe.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>