labs/sysdev-realtime: remove pointless rttest directory
For some reason, the data of the real-time lab are not directly under
realtime/, but under realtime/rttest/, with realtime/ containing nothing
but the rttest/ directory. This is rather pointless and inconsistent
with the organization of files in other labs.

Therefore, we move everything to realtime/, and adjust the lab
instructions accordingly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>