slides/kernel-frameworks2: improve device-managed allocation slides
They were added in the middle of a completely unrelated explanation
(links between structures), which was very confusing. Plus the
explanation about them wasn't really generic (it was really tied to
the very specific topic being presented in the surrounding slides).

So instead, move it earlier, in a separate sub-section, with a more
generic explanation. An actual example is added to clearly show the
benefit of such functions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>