crosstool-ng lab: workaround to fix copy-paste on apt-get line
These lines are rendered in full-sysdev-labs.pdf as two
columns. There's no visual difference, but copy-pasting them as the
lab book suggests does not work. The pasted text looks like:

  sudo apt-get install
  libncurses5-dev
  build-essential
  pkg-config wget
  autoconf automake libtool-bin libexpat1-dev \
  bison flex patch curl cvs texinfo git bc \
  subversion gawk python-dev gperf unzip \
  help2man

In place of:

  sudo apt-get install autoconf automake libtool-bin libexpat1-dev \
  libncurses5-dev bison flex patch curl cvs texinfo git bc \
  build-essential subversion gawk python-dev gperf unzip \
  pkg-config help2man wget

The issue is triggered by the fact that all lines have a whitespace in
the same column. I'm unable to fix the problem at its root, but
swapping two packages is an effective workaround.

Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>