labs/sysdev-application-debugging: don't copy the entire Buildroot
Copying the entire Buildroot folder with all the build output and then
cleaning it up is really silly. Ideally, we should have people use out
of tree build, but that's probably a bit too confusing for a sysdev
toolchain aimed at beginners.

So instead, ask participants to re-extract the Buildroot source code,
and copy the Buildroot configuration from the previous build.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>