labs/sysdev-flash-filesystems: change how to get the LEB size
Running ubiattach on an unitialized partition, while we haven't yet
done the partitioning of our flash, just for the sake of getting the
LEB size, seems a bit weird.

It seems a lot more logical to look at the NAND flash characteristics,
and use the explanations from the slides that allow to understand how
the LEB size is calculated from the PEB size, the page size and
sub-page size.

So we use "nand info" in U-Boot, and then explain how to convert those
details into the LEB size.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>