yocto-first-build: Remove the build directory's creation into poky
This is not the first time that I got some trainees that do not
understand (and get confused) when I repeat "do not edit poky"
and that they need to edit the configuration of the "build" folder
(because they see it under "poky" folder).

Moreover, personally, I like to have the build folder in the same
level than layers.

Now, we will have:
     build/
     meta-felabs/
     meta-ti/
     poky/
instead of
     meta-felabs/
     meta-ti/
     poky/build/

which is more logical to me.

Alex, Antoine, if you disagree with me, tell me why and it is
still possible to revert this commit.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>