yocto: lab-data: Create a subfolder "bootlin-lab-data"
To have a better organization with layers and scripts/patches
we are providing, create a subfolder "bootlin-lab-data" on
the tarball yocto-labs.tgz.
Thanks to that, the yocto-labs folder will look like:
       .
       | bootlin-lab-data
       | build
       | meta-felabs
       | meta-ti
       | poky

instead of:
	.
	| 0001-Simplify-linux-ti-staging-recipe.patch <-- from our tarball
	| build
	| meta-felabs
	| meta-ti
	| nunchuck <-- from our tarball
	| poky
	| script <-- from our tarball

The previous organization can mix up trainees that are not familiar with
Yocto Project and get confused with the "script" folder, for example.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>