kernel-driver-development-debugging: adjust explanation about dynamic debug
Some people where confused about the behavior of
DEBUG/CONFIG_DYNAMIC_DEBUG: it wasn't clear to them that it's only
when neither DEBUG nor CONFIG_DYNAMIC_DEBUG are used that the messages
are not compiled in. I.e, they thought they needed both the debugging
messages to be enabled.

Let's reword a bit the explanation in an attempt to clarify that.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>