yocto: update to sumo
sumo is the 2.5 release.
poky for sumo is release 19.0.0.
updated meta-ti to latest tag
meta-ti is now using linux-ti-staging_14.4 (and not 4.9) and a newer
U-Boot.

yocto-layer has been replaced by bitbake-layers create-layer and
yocto-bsp has been removed.

LAYERSERIES_COMPAT is now required in layer.conf to tell with which
Yocto release it's compatible.

meta-ti layer needs a bit more tweaking to work:
 - base_read_file produces warnings in bitbake so they have been
 replaced with the oe core function,
 - U-Boot from TI does not work since 2018.00-rc3, revert the patch that
 breaks the BeagleBone Black,
 - add LAYERSERIES_COMPAT since it's now required else we get a warning,
 - Replaced kernel with ${KERNEL_PACKAGE_NAME} in the linux recipe since
 it made bitbake display a few warnings,

Signed-off-by: Quentin Schulz <quentin.schulz@bootlin.com>