labs: yocto: try to give a better hint that something's wrong
I've seen trainees not adding the bootloader to the machine either
because they didn't read that part or didn't take the hint.

Reformulate the title of that part so it's a bit more explicit what has
to be done (and it is redundant with the question so hopefully they'll
get it).

Signed-off-by: Quentin Schulz <quentin.schulz@bootlin.com>