Properly reduce the agenda PDF size
The issue was that header.svg included a huge bitmap. The default
logic of our build system is to render .svg into .pdf files, assuming
they contain only vector graphics. This makes the header.pdf file
particularly large, and therefore any PDF including it is equally
large.

The same bitmap in .png is also quite large, due to the background
picture that doesn't compress well. We really need the background
picture to be in .jpg format. However, having the logos in .jpg would
mean having artifacts.

So instead, what we do is that we combine the best of both worlds and
do the composition directly in LaTeX: the background is in .jpg
format, and we overlay the two logos in .pdf on top of it.

It requires yet another LaTeX package, the "overpic" package.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>