labs/kernel-i2c-input-interface: improve the wording
The "First, add lines [...]" text was very confusing, because the
explanation about the private structures is so long that we no longer
remember that we were working on the nunchuk_poll() function. And in
fact, just a few lines above, we are working on nunchuk_probe(), which
is very confusing.

Try to improve the wording to make it clear we are now back to
implementing the nunchuk_poll() function.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>