Makefile: fixup generation of single chapter slides
When running "make sysdev-realtime-slides.pdf", one expects to build
just the realtime chapter. However, this was no longer working.

Indeed, we extract the prefix, here "sysdev" to determine the training
course. But in fact the variable containing the list of all slides for
the embedded Linux course is EMBEDDED_LINUX_SLIDES, not SYSDEV_SLIDES,
so we need to accomodate for that.

Also, the UPPERCASE macro needs to convert "-" to "_", so that
lowercase embedded-linux because EMBEDDED_LINUX in uppercase.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>