Kernel labs: use a github.com link
People in the training (including myself) had a hard time finding the
pre-compiled binaries because linking to the raw README.txt file
points to a raw.githubusercontent.com link, from which it is
"difficult" to find the binaries listed in the file. Point to the
github.com page instead, people will find out the README.TXT by
themselves (it is displayed on the webpage pointed anyway).

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>