Flashing BBB: don't lie about the wait delays
It took less than 10 seconds on my BBBW to get all the LEDs
blinking. It happened so fast compared to the lab instructions that I
was pretty sure something went wrong and I retried it a couple of
times. Let's not lie too much about the delay to avoid these
situations.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>