agenda: rework Embedded Linux agenda
The Embedded Linux agenda, which is postponing the solutions of the
labs each time to the next day doesn't work very well. For the
participants, seeing the demos right after the lectures if much more
important.

So switch back to a model where lectures and demos are interleaved, as
they used to be. If participants wants to repeat the demos using the
optional labs, they can still do so between the sessions.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>