kernel: serial-iomem:
Assuming TTL UART, the Tx line is an output and the Rx line is an input:
* The Rx line must remain stable at a high logic level if the other end
  gets disconnected, hence the use of a an internal pull-up makes
  totally sense.
* However, while configured in UART mode, the Tx line will be driven, so
  we don't really care about setting an internal pull-up or pull-down.

The current pull-down configuration was copy-pasted from other device
trees, which probably were also copy-pasted from the work of a single
person which only focused on the fact that it worked configured like
this (and it does, the additional wired resistor does nothing on a
driven output). So let's get rid of it.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>