kernel: workqueue: Mention the default queue
In the slide we talked about queue_work() (for queueing a work in a
specific queue) and used schedule_work() in the example without
mentioning its purpose (the use of a default queue). Add a mention to it
in the explanation.

Reduce a little bit the size of the minted block in order to fit in the
slide.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>