kernel: sources: Rework the slide about in-tree drivers
This slide is more advertising the benefits of mainlining code more than
organizing code. "in-tree" drivers do not mean that they are upstream
(that was the purpose of one of the previous slides). Here we are
talking about mainline code so rename the slide. Also rework a little
bit the list of advantages, getting contributions from customers is
anecdotal, getting reviews is actually useful and should be seen
positively, a mainline driver will not suffer from API changes and will
be compatible with all the kernel versions onward.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>