sysdev/kernel: building: Fully rework the slide about the cmdline
The original purpose of this slide was to explain how the kernel was
getting its cmdline.

While explaining that the kernel configuration has an impact on the
cmdline might be worth keeping, I do not think it is wise to talk about
device tree overlays, U-Boot fixing the DT and so on. In particular
because device tree will be addressed later in the training(s). Here, we
are just explaining how to boot the kernel with U-Boot, nothing more,
it's important to keep this portion short and clear, so let's not bother
with too much detail. Plus, U-boot natively does noting of all of that,
it's usually a user input/script more than a core/automated feature.

The reference to DT fixups is the slide title is dropped, if fixups and
overlays need to be explained, this should be a dedicatd slide among the
existing device tree slides comming next.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>