kernel: i2c: Clarify why we should use the SMBus protocol
Using SMBus commands is not mandatory but preferred (see the kdoc) and
only when talking to an SMBus device, so make this clear. While at it,
clarify the reason for that choice: SMBus commands will be translated
into I2C commands by the core if needed, but the opposite is not
possible.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>