kernel: i2c: Fix the SMBus example
I2C/SMBus do not support two Start bits in the same sequence. It is
actually a repeated start (abbreviated Sr) which must be performed after
the command byte to signify that we change the direction of the flow:
the host wrote the command, now it's the peripheral responsibility to
talk on the bus.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>