yocto: adapt to cover the ST32MP157A variant
The A variant is still useful as the D is hard to find, so better be able
to use any of the two.

The 157A is almost compatible with the 157D, so this is mostly a matter of
just mentioning it where applicable and to re-add and update the nunchuk DT
patch removed in 0ec7575b62a157c0814c7c84dd5dbad73276df0b.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>