yocto: slides: improve PREFERRED_VERSION example
Python 2 is really EOL now and it's not in any layer we use in the
examples. Now there is only Python 3, the recipe is called "python3" but it
implements one version only. For a more realistic example, switch to a
package that currently has at least two versions. Unfortunately oe-core has
no such recipe, thus pick one from meta-openembedded, which is nginx having
the "stable" and "current" versions.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>