yocto: slides: move varflags slide before varflags are used (almost)
Explaining PACKAGECONFIG is quite hard, not only because it is an advanced
topic, but also because it is heavily based on varflags, which are
introduced later on.

Move the "Variable flags" subsection before PACKAGECONFIG (still within the
"Writing recipes - going further" section).

Also move yocto-licensing after yocto-recipe-extra as it uses varflags too,
and there is no cross-dependency between those two sections.

According to my search, there is no other use of varflags in the slides
except for SRC_URI[*sum], but let's keep it as is: SRC_URI is such a basic
variable that it's not possible to move it late in the slides, and moving
varflags earlier than that would anticipate a slightly complex topic for no
reason.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>