yocto: slides: improve varflags examples
Add a SRC_URI[md5sum] example for two reasons: make it clear that varflags
apply also to variables (not only tasks) and as a hint for the speaker to
mention "we've already seen a varflag usage".

Change the "nostamp" example to something where "nostamp" makes sense, such
as do_menuconfig.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>