yocto: slides: add overview picture showing binary packages
Showing early how yocto builds binary packages and not directly a rootfs is
expacted to remove a lot of confusion in trainees. Do it with a modified
copy of the buildsystem-principle.dia picture.

Also add a link to the big and complex picture in the Yocto
documentation. The idea is that the trainer can click to open this picture
and keep it open during the training to describe the blocks one by one when
the corresponding topic is described in the training.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>