labs/sysdev-thirdparty: fix alsa-lib instructions
First, the ALSA config files should be installed as part of the
alsa-lib step, not the alsa-utils step, because they are part of
alsa-lib, not alsa-utils.

Second, instead of cherry-picking a random set of config files, which
leads to a "speaker-test" working but still spitting out a bunch of
error messages, we copy the full set of configuration files (simpler)
and adjust the one thing that causes problem. We now have simpler
instructions, and a speaker-test that doesn't spit out weird error
messages.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>