nunchuk-mpd-client: be less strict on the expected joystick ABS_X/ABS_Y values
With my Nunchuk, I never had a 0 or 255 value in ABS_X, the min I
could get was 1, the max was 254. To accomodate for this, be a bit
more flexible on the values that trigger a song next/prev. For
consistency, we do the same for ABS_Y for volume up/down.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>