sysdev: slides: networking-stack.dia: swap socket/netlink, improve block alignment
Rationale:

 * The various servers/clients applications (right side) mostly use the
   socket API (left side) to use TCP/UDP/IP (right side).
 * The various management/config applications (left side) go in large part
   via the netlink interface (right side) to reach assorted thing,
   including firewall/qos (left side).

So let's have the two swapped to have less conceptual "crossing" between
the two halves of the picture. This is not a perfect split as netlink goes
through the socket interface, but a perfect picture would be poorly
readable, and I would not be able to draw it anyway. :)

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>