kernel: concurrency: Improve the first spinlock slide
Change a little bit the text of the picture just to mention that with
preemption we may end up in a deadlock, but this is not 100%
guaranteed. Of course, not disabling preemption is a bug and should be
considered very bad practice, the goal is just to be a little bit more
accurate, more details can be given orally if the audience already has
an advanced understanding of the problem.

While on this slide, update a bit the text to reflect where we use it as
well (not only in process context, that's the whole point).

Also mention the prevention of the migration which is implied when
disabling the preemption.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>