kernel: Introduce a real DT chapter
This is still work in progress, the chapter for the sysdev training has
been pulled, the first explanation of the device tree adapted to the
BBB, I need to update a few other examples to better match the purpose
of the kernel training.

Then, we will have to create a new small lab inviting people to play
with the device tree, followed by a lecture explaining the parsing of
the device tree and what the kernel does to create devices internally.

With all this, we will eventually will drop the "equivalent" chapter
about DT, including the platform device explanation.

TODO: The pinmuxing part has been commented out for now, it will be
updated later.
TODO: We need to do a first lab at the end of this chapter.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>