slides: add provision for an "Extra section" in appendix
For the sysdev training we want to create an "Extra slides" section as an
appendix. To avoid them to be counted in the total slide number, use the
appendixnumberbeamer package and the needed commands at the end of
last-slides. These changes have no effect unless any tex file is added
after last-slides.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>