sysdev: slides: linux versioning: remove info about release candidates
As agreed with Thomas and Michael, mentioning rc versions in the sysdev
training is not very important and adds too much details so let's remove
it from the picture.

The kernel training already has a copy of this same picture, but it is OK
to leave rc mentioned there so let's leave it as is.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>