embedded-linux: on stm32, don't load kernel right at start of RAM
That's not recommended practise, as the compressed kernel would
overlap with the uncompressed kernel, as explained on
https://people.kernel.org/linusw/how-the-arm32-linux-kernel-decompresses

As a consequence, the kernel bootstrap code has to make a copy of the
compressed kernel far enough in RAM so there is no such overlap.

Tests on stm32mp157a-dk1 show a 107 ms boot time reduction
(observed consistently on 2 series of 5 tests).

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>