sysdev: labs: consolidate enabling ltrace, strace and perf at the beginning of appdev/debugging lab
Building ltrace and strace in the systemd lab is not intuitive, as they are
not used there. So move it in the appdev/debugging lab, just before using
them. Also move ahead enabling perf so that only one rebuild is needed
during this lab.

Note this does not affect the qemu labs, which have no systemd lab and have
separate instructions for application development and debugging.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>