Makefile: copy symbolic links when creating lab-data archive
This way, we can have symbolic links in the sources,
to share files between several lab variants (such as
"embedded-linux", "embedded-linux-bbb" and "embedded-linux-qemu").

That's better than keeping multiple copies of the same file
in the sources, which may go out-of-sync if a contributor
doesn't notice they are supposed to be the same.

Tested on the embedded-linux-bbb-labs.tar.xz archive.

Signed-off-by: Michael Opdenacker <michael.opdenacker@bootlin.com>
Reported-by: Amir Farahani Khojasteh <amfa00001@stud.uni-saarland.de>