slides/kernel-driver-development-io-memory: Clarify a list of functions
People not so used to the command line do not understand very well
read[bwlq] and interpret that as an array. As these functions are
actually quite fundamental, let's list them all one by one, it does not
hurt, so that later uses of read[bwlq] are actually understood the right
way.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>