debugging: labs: add workaround for broken perf uprobe listing
When running the perf list command on target to get available sites to
insert uprobes, the output may be quite broken because perf is trying to
colorize the output with ANSI escape sequences, which may be not properly
interpreted depending on the setup. This res:

[[m[[34m         [[m[[34m
[[m     32  	        clock_gettime(CLOCK_MONOTONIC, &start);
     33  	        start_nano = timespec_to_nano(&start);

This makes teh research for useful info (here, a line number) quite
difficult.

I have tried the following things, without success:
* change host terminal
* change serial communication program
* change TERM variable on system
* check if buildroot has options to configure terminal
* check if perf has options to disable output colorization