linux-kernel: Fix sysdev vs. kernel split regarding the versioning
The common chapter was duplicated in order for the sysdev slides to be
simplified, but the Makefiles were not updated, leading to a simplified
version in the kernel as well. I actually think we can simplify this a
bit as well, so let's start by at least using this duplicated chapter.

Fixes: 594d6aff ("slides/linux-kernel-intro-versioning: new section for the kernel course")
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>