kernel: intro: sources: Don't mention realtime as sub-community dev tree
Mentionning real-time in the list of sub-community treeis is misleading,
as we state these trees aren't meant to be used in products.

Realtime is a bit different, as there are stable trees with RT that
are maintained, containing both stable fixes from linux-stable and
fixes to the RT patch.

Instead, mention other dev trees for core subsystems.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>