debugging: labs: replace $< with $^ in system_profiling lab Makefile
While both $< and $^ are valid to manipulate a single prerequisite in a
makefile target, it can lead to silent issue when updating the makefile.
For example, we need to add the tracepoint provider generated object at
some point in the final binary, and it is very likely that both trainer and
trainees will forget to update this symbol to take all prerequisites
instead of only the first one, if they do not copy-paste directly the
makefile sample given in labs instruction

Prevent this omision by using directly $^ in base Makefile for system
profiling lab.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>