slides: yocto: remove the usage of PRIORITY in recipes
PRIORITY is not to be used within recipe themselves according to the
documentation:

"PRIORITY is considered to be part of the distribution policy because
the importance of any given recipe depends on the purpose for which the
distribution is being produced. Thus, PRIORITY is not normally set
within recipes."

Present SUMMARY as it is widely used in openembedded-core and
meta-openemebedded.

Signed-off-by: Charles Perry <charles.perry@savoirfairelinux.com>