debugging: slides: kernel: reorder slides, add subsections
Rearrange the slides order in a more rational way, starting from good
practices to avoid bugs to progressively more complex and powerful
debugging mechanisms.

Split the section in smaller subsections to provide a high-level overview
of where each topic fits in the big picture, both for trainees and for the
trainer.

Discussed-with: Alexis Lothoré <alexis.lothore@bootlin.com>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>