yocto: labs: workaround for missing space before backslash in bashinput
bashinput incrrectly renders without a space before the backslash, which
results in:

| sudo apt install bc build-essential chrpath cpio diffstat gawk git python3 texinfo wget gdisk\
|     libssl-dev lz4

the missing space between "gdisk" and "" implies that copy-pasting these
lines in a terminal results in:

  E: Unable to locate package gdisklibssl-dev

This appears as a bug in the bashinput environment word wrapping. Use the
verbatim environment until a suitable fix for bashinput is implemented.

Report-at: https://github.com/bootlin/training-materials/issues/221

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>