yocto: slides: basics: split "The build directory" in initial and post-build
Mention only the 'conf' subdir initially, and also move it a bit before to
be closer to the oe-init-build-env slide. Move the other subdirectoried
after the 'Compilation' slide, when these subdirs will exist.

This new layout can also help trainers who need to start the build as early
as possible.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>