kernel-beagleplay: Add serial labs
The main change in the serial labs is that we use UART5 and UART6 instead of
UART2 and UART3. This changes the instructions for the pinmuxing section of the
labs.

The DMA lab has some big differences with the original version, as the AM62x
does not use the same DMA controller as the AM335x. The controller used by the
AM62x does not have the "first byte" quirk so the driver needs to be adjusted
to account for this.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>