debugging: labs: specify lttng-relayd output path out of nfs root
Trainees may encounter issues with the LTTNG lab if they set the ouput
directory of the relay inside the rootfs exposed through NFS. A training
session allowed us to observe the issue, not right at relay's start but
whenever we try, from the target, to enable a specific event. When such
operation is done, the following error appears on the host:

user@host:$ lttng-relayd --output=$PWD/traces
PERROR - 14:18:20.065382791 [30341/30345]: mkdirat recursive last element: Permission denied (in create_directory_recursive() at directory-handle.c:986)
Error: Failed to create session output path "[...]/debugging-labs/nfsroot/root/system_profiling/traces/buildroot/crc_session-20000101-054200"

Based on the error (permission denied on mkdirat), the issue is likely due
to a missing write permission on nfs directory (could be linked to user
mismatch too between host and target). Tests showed that the error
disappear if we provide a path out of the nfsroot.

In order to avoid such issues, which can go silent if we do not check back
the relay's logs, ask specifically to use a path out of the nfsroot.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>