kernel-beagleplay: Fall back to booting from SD card
Ideally, we would've prepared instructions for booting from eMMC using
Snagboot but the configs and U-Boot versions suitable for recovering the
board have broken networking drivers. Another possible solution would've
been to recover the board with recent images and flash the older sysdev
images to eMMC, since networking works with these. The problem with this is
that these images save the environment to the SD card. Why not rebuild them
then? Well, there's a bug in recent versions of OpenSSL that breaks the old
sysdev build.

At this, point this whole issue is getting too complicated to fix properly
so just fall back to booting from an SD card.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>