driver-architecture: bus infrastructure is itself a framework
This is explained much later in the training, but I think it's important
to mention this early on already. Otherwise, things that are explained
immediately afterwards don't make much sense (e.g. the fact that you
have a struct device twice: once for the bus and once for the functional
framewoek).

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>