sysdev-kernel: building: Stop mentioning legacy C strucs too early
We now have an extended chapter about the DT, including the other means
(ACPI and legacy C structures) in both trainings. We need to mention
here the need for giving a DTB to the kernel, but in practice there is
no need to detail more, because every time I feel like I need to tell
more about all of that and I later repeat myself again.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>