kernel: dt: Reorder some slides
It probably makes more sense to talk about the properties (the content
of the DT) before explaining the bindings (how to check the
content). Let's move the few slides talking about this *after* the
section explaining the properties and their meanings. Once we know the
common properties, we can use them in the bindings.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>