sysdev/kernel: hw-devices: Save again the drawing
On Ubuntu 22.04/23.10 and Dia 0.97.3 the drawing looks fine.
On Fedora with the same Dia version the drawing looks bery bad.
Opening the file and saving it leads to changes in the file content
(even though I didn't modify the drawing at all). Let's try it and see
if it fixes the rendering.

Reported-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>