labs: sysdev-toolchain: Explicitly select hardware FPU
If we leave the 'auto (let gcc decide)' as floating point configuration,
the softare FPU is selected so the toolchain is installed in
x-tools/arm-training-linux-musleabi while the rest of the labs expects
it to be installed in x-tools/arm-training-linux-musleabihf.

Add an explicit selection of the hardware FPU in the ct-ng menuconfig.

Signed-off-by: Bastien Curutchet <bastien.curutchet@bootlin.com>