debugging: slides: move libpcap internals details to eBPF section
While discussing how libpcap processes user filters into bpf programs is
very useful to introduce bpf/ebpf, it may be too advanced for the section
where it is currently located (common debugging tools), and so feel a
little bit off topic.

Move it to the eBPF section as an introduction. While at it, fix the
userspace<->kernel interface: original BPF does not use the bpf() syscall
interface, but the socket()/setsockopt() interface. That makes another
reason to use it as an introduction to extended BPF rather than an
example.

Signed-off-by: Alexis Lothoré <alexis.lothore@bootlin.com>